fix(logs): change size of bytes sent

+++ src/main/java/backend/academy/project/logs/LogRecord.java
@@ -11,7 +11,7 @@ public record LogRecord(
      String requestedResource,
      String httpVersion,
      int status,
-     int bytesSent,
+     long bytesSent,
      String httpReferer,
      String httpUserAgent
     )
